+++ src/app/[locale]/account/(withoutLeftNav)/profile/data-access-requests/applications/[applicationId]/components/ApplicationSection.tsx
@@ -32,7 +32,6 @@ import useDelete from "@/hooks/useDelete";
 import useDialog from "@/hooks/useDialog";
 import notificationService from "@/services/notification";
 import apis from "@/config/apis";
-import config from "@/config/config";
 import { inputComponents } from "@/config/forms";
 import {
     beforeYouBeginFormFields,
@@ -48,10 +47,10 @@ import {
     ARRAY_PREFIX,
     DarApplicationApprovalStatus,
     DarApplicationStatus,
+    DarTemplateType,
 } from "@/consts/dataAccess";
 import { ArrowBackIosNewIcon, HelpOutlineIcon } from "@/consts/icons";
 import { RouteName } from "@/consts/routeName";
-import { setTemporaryCookie } from "@/utils/cookies";
 import {
     createFileUploadConfig,
     formatDarAnswers,
@@ -253,12 +252,6 @@ const ApplicationSection = ({
             );
 
         if (formData) {
-            setTemporaryCookie(
-                config.DAR_UPDATE_SUPPRESS_COOKIE,
-                Date.now().toString(),
-                60
-            );
-
             const [resAnswers, resApplication] = await Promise.all([
                 updateDarApplicationAnswersAction(applicationId, userId, {
                     ...applicationData,
@@ -332,6 +325,36 @@ const ApplicationSection = ({
         ? parentSections.findIndex(section => section.id === sectionId)
         : 0;
 
+    const formatFileUploadFields = (component: string, questionId: number) => {
+        let fileUploadFields: FileUploadFields | undefined;
+
+        if (
+            component === inputComponents.FileUpload ||
+            component === inputComponents.FileUploadMultiple ||
+            component === inputComponents.DocumentExchange
+        ) {
+            const fileDownloadApiPath = isResearcher
+                ? `${apis.usersV1Url}/${userId}/dar/applications/${applicationId}/files`
+                : `${apis.teamsV1Url}/${teamId}/dar/applications/${applicationId}/files`;
+
+            fileUploadFields = createFileUploadConfig(
+                questionId.toString(),
+                component,
+                applicationId,
+                fileDownloadApiPath,
+                isResearcher,
+                setValue,
+                getValues,
+                teamApplication?.submission_status !==
+                    DarApplicationStatus.SUBMITTED
+                    ? removeUploadedFile
+                    : undefined
+            );
+
+            return fileUploadFields;
+        }
+    };
+
     const renderFormFields = () =>
         filteredData
             ?.filter(
@@ -344,34 +367,12 @@ const ApplicationSection = ({
                 if (field.is_child) return null;
 
                 let sectionHeader = null;
-                if (!processedSections.has(field.section_id)) {
+                if (
+                    !processedSections.has(field.section_id) &&
+                    data.application_type !== DarTemplateType.DOCUMENT
+                ) {
                     processedSections.add(field.section_id);
                     sectionHeader = renderSectionHeader(field);
-                }
-
-                let fileUploadFields: FileUploadFields | undefined;
-
-                if (
-                    field.component === inputComponents.FileUpload ||
-                    field.component === inputComponents.FileUploadMultiple
-                ) {
-                    const fileDownloadApiPath = isResearcher
-                        ? `${apis.usersV1Url}/${userId}/dar/applications/${applicationId}/files`
-                        : `${apis.teamsV1Url}/${teamId}/dar/applications/${applicationId}/files`;
-
-                    fileUploadFields = createFileUploadConfig(
-                        field.question_id.toString(),
-                        field.component,
-                        applicationId,
-                        fileDownloadApiPath,
-                        isResearcher,
-                        setValue,
-                        getValues,
-                        teamApplication?.submission_status !==
-                            DarApplicationStatus.SUBMITTED
-                            ? removeUploadedFile
-                            : undefined
-                    );
                 }
 
                 return (
@@ -418,8 +419,15 @@ const ApplicationSection = ({
                                         },
                                         control,
                                         field.question_id.toString(),
-                                        updateGuidanceText,
-                                        fileUploadFields
+                                        field.component !==
+                                            inputComponents.DocumentExchange &&
+                                            updateGuidanceText,
+                                        field.component &&
+                                            formatFileUploadFields(
+                                                field.component,
+                                                field.question_id
+                                            ),
+                                        field.document
                                     )}
                                 </Box>
                             </>
@@ -456,7 +464,12 @@ const ApplicationSection = ({
                                                     },
                                                     control,
                                                     child.question_id.toString(),
-                                                    updateGuidanceText
+                                                    updateGuidanceText,
+                                                    child.component &&
+                                                        formatFileUploadFields(
+                                                            child.component,
+                                                            child.question_id
+                                                        )
                                                 )}
                                             </Box>
                                         </Fragment>
@@ -577,10 +590,10 @@ const ApplicationSection = ({
                                 <Typography
                                     variant="h2"
                                     sx={{ p: 2, pl: 0, pb: 1 }}>
-                                    {sections[sectionId].name}
+                                    {sections[sectionId]?.name}
                                 </Typography>
                                 <Typography>
-                                    {sections[sectionId].description}
+                                    {sections[sectionId]?.description}
                                 </Typography>
                             </Box>
                             <Divider variant="fullWidth" />
@@ -664,49 +677,55 @@ const ApplicationSection = ({
 
                         {parentSections.find(
                             section => section.id === sectionId
-                        )?.name !== messageSection.name && (
-                            <Box sx={{ flex: 1, overflowY: "auto", p: 0 }}>
-                                <Box>
-                                    <Typography
-                                        variant="h3"
-                                        sx={{
-                                            display: "flex",
-                                            alignItems: "center",
-                                            justifyContent: "center",
-                                            m: 0,
-                                        }}>
-                                        <HelpOutlineIcon
-                                            sx={{
-                                                mr: 1,
-                                                color: colors.grey600,
-                                                fontSize: 16,
-                                            }}
-                                        />
-                                        {t("guidance")}
-                                    </Typography>
-                                </Box>
-                                <Divider variant="fullWidth" sx={{ mb: 4 }} />
-                                <Box
-                                    sx={{
-                                        pt: 0,
-                                        pb: 0,
-                                    }}>
-                                    {guidanceText ? (
-                                        <MarkDownSanitizedWithHtml
-                                            content={guidanceText}
-                                        />
-                                    ) : (
+                        )?.name !== messageSection.name &&
+                            data.application_type !==
+                                DarTemplateType.DOCUMENT && (
+                                <Box sx={{ flex: 1, overflowY: "auto", p: 0 }}>
+                                    <Box>
                                         <Typography
+                                            variant="h3"
                                             sx={{
-                                                color: theme.palette.grey[500],
-                                                textAlign: "center",
+                                                display: "flex",
+                                                alignItems: "center",
+                                                justifyContent: "center",
+                                                m: 0,
                                             }}>
-                                            {t("defaultGuidance")}
+                                            <HelpOutlineIcon
+                                                sx={{
+                                                    mr: 1,
+                                                    color: colors.grey600,
+                                                    fontSize: 16,
+                                                }}
+                                            />
+                                            {t("guidance")}
                                         </Typography>
-                                    )}
+                                    </Box>
+                                    <Divider
+                                        variant="fullWidth"
+                                        sx={{ mb: 4 }}
+                                    />
+                                    <Box
+                                        sx={{
+                                            pt: 0,
+                                            pb: 0,
+                                        }}>
+                                        {guidanceText ? (
+                                            <MarkDownSanitizedWithHtml
+                                                content={guidanceText}
+                                            />
+                                        ) : (
+                                            <Typography
+                                                sx={{
+                                                    color: theme.palette
+                                                        .grey[500],
+                                                    textAlign: "center",
+                                                }}>
+                                                {t("defaultGuidance")}
+                                            </Typography>
+                                        )}
+                                    </Box>
                                 </Box>
-                            </Box>
-                        )}
+                            )}
                     </Box>
                 </Paper>
             </Box>