+++ src/app/actions/revalidateCacheAction.ts
@@ -3,5 +3,5 @@
 import { revalidateCache } from "@/utils/revalidateCache";
 
 export const revalidateCacheAction = async (tags: string[] | string) => {
-    revalidateCache(tags);
+    revalidateCache(tags, true);
 };