+++ src/components/MarkDownSanitizedWithHTML/index.tsx
@@ -62,7 +62,31 @@ export const MarkDownSanitizedWithHtml = ({
                 component: CustomLink,
             },
         }),
-        p: <Typography sx={{ mb: 2 }} />,
+        p: {
+            component: Typography,
+            props: { sx: { mb: 2 } },
+        },
+        h2: {
+            component: Typography,
+            props: {
+                variant: "h2",
+                sx: { fontWeight: 600 },
+            },
+        },
+        h3: {
+            component: Typography,
+            props: {
+                variant: "h3",
+                sx: { fontWeight: 600 },
+            },
+        },
+        h4: {
+            component: Typography,
+            props: {
+                variant: "h4",
+                sx: { fontWeight: 600 },
+            },
+        },
     };
 
     const Wrapper = wrapper as React.ElementType;