+++ src/utils/revalidateCache.ts
@@ -1,13 +1,18 @@
 import { refresh, revalidateTag } from "next/cache";
 
-export const revalidateCache = (tags: string[] | string) => {
+export const revalidateCache = (
+    tags: string[] | string,
+    refreshCache?: boolean
+) => {
     if (Array.isArray(tags)) {
         tags.forEach(tag => {
             revalidateTag(tag, "max");
-            refresh();
         });
     } else {
         revalidateTag(tags, "max");
+    }
+
+    if (refreshCache) {
         refresh();
     }
 };